// src/components/MapPartComponents/data.d.ts
export interface AgvInfoState {
        agvId: string,
        agvCode?: string | number,
        agvName:string,
        agvStatus:string,
        missionId?: string | number,
        missionName?: string | number,
        agvVoltageLow: number,
        agvVoltageFull: number,
        voltage: number,
        startCode?:number | string,
        endCode?:number | string,
}

export const agvInfoInit={
    agvId: "",
    agvCode : "",
    agvName: "",
    agvStatus: "",
    missionId: "",
    missionName: "空闲",
    agvVoltageLow: 0,
    agvVoltageFull: 0,
    voltage: 0,
    startCode:'',
    endCode:''
}